+++ js/app/components/delivery-form/DeliveryForm.js
@@ -26,7 +26,7 @@ import Price from './Price'
 import SuggestionModal from './SuggestionModal'
 import DeliveryResume from './DeliveryResume'
 import Map from '../DeliveryMap'
-import { Mode } from './Mode'
+import { Mode } from './mode'
 import { useSelector } from 'react-redux'
 import { selectMode } from './redux/formSlice'
