+++ js/app/components/delivery-form/Price.js
@@ -12,7 +12,7 @@ import { useDeliveryFormFormikContext } from './hooks/useDeliveryFormFormikConte
 import _ from 'lodash'
 import OverridePriceForm from './OverridePriceForm'
 import { useCalculatePriceMutation, useGetTaxRatesQuery } from '../../api/slice'
-import { Mode } from './Mode'
+import { Mode } from './mode'
 import { useSelector } from 'react-redux'
 import { selectMode } from './redux/formSlice'
